+++ mockserver-core/src/main/java/org/mockserver/matchers/XPathMatcher.java
@@ -4,11 +4,17 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.apache.commons.lang3.StringUtils;
 import org.mockserver.log.model.LogEntry;
 import org.mockserver.logging.MockServerLogger;
+import org.mockserver.xml.StringToXmlDocumentParser;
+import org.mockserver.xml.XPathEvaluator;
+import org.xml.sax.SAXException;
 
+import java.io.IOException;
 import java.util.Iterator;
 import java.util.Map;
 import javax.xml.XMLConstants;
 import javax.xml.namespace.NamespaceContext;
+import javax.xml.namespace.QName;
+import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.xpath.XPath;
 import javax.xml.xpath.XPathConstants;
 import javax.xml.xpath.XPathExpression;
@@ -16,20 +22,17 @@ import javax.xml.xpath.XPathExpressionException;
 import javax.xml.xpath.XPathFactory;
 
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
-import static org.mockserver.matchers.StringToXmlDocumentParser.ErrorLevel.prettyPrint;
+import static org.mockserver.xml.StringToXmlDocumentParser.ErrorLevel.prettyPrint;
 import static org.slf4j.event.Level.DEBUG;
-
 
 /**
  * @author jamesdbloom
  */
 public class XPathMatcher extends BodyMatcher<String> {
-    private static final String[] EXCLUDED_FIELDS = {"mockServerLogger", "stringToXmlDocumentParser", "xpathExpression"};
+    private static final String[] EXCLUDED_FIELDS = {"mockServerLogger", "xPathEvaluator"};
     private final MockServerLogger mockServerLogger;
     private final String matcher;
-    private final boolean namespaceAware;
-    private final StringToXmlDocumentParser stringToXmlDocumentParser = new StringToXmlDocumentParser();
-    private XPathExpression xpathExpression = null;
+    private XPathEvaluator xPathEvaluator = null;
     XPathMatcher(MockServerLogger mockServerLogger, String matcher) {
         this(mockServerLogger, matcher, null);
     }
@@ -37,40 +40,17 @@ public class XPathMatcher extends BodyMatcher<String> {
     XPathMatcher(MockServerLogger mockServerLogger, String matcher, Map<String, String> namespacePrefixes) {
         this.mockServerLogger = mockServerLogger;
         this.matcher = matcher;
-        this.namespaceAware = namespacePrefixes != null;
         if (isNotBlank(matcher)) {
             try {
-              XPath xpath = XPathFactory.newInstance().newXPath();
-              if (namespacePrefixes != null) {
-                xpath.setNamespaceContext(new NamespaceContext(){
-                  public String getNamespaceURI(String prefix) {
-                    if (namespacePrefixes.containsKey(prefix)) {
-                      return namespacePrefixes.get(prefix);
-                    }
-                    return XMLConstants.NULL_NS_URI;
-                  }
-              
-                  // This method isn't necessary for XPath processing.
-                  public String getPrefix(String uri) {
-                      throw new UnsupportedOperationException();
-                  }
-              
-                  // This method isn't necessary for XPath processing either.
-                  public Iterator getPrefixes(String uri) {
-                      throw new UnsupportedOperationException();
-                  }
-                });
-              }
-              
-              xpathExpression = xpath.compile(matcher);
-            } catch (XPathExpressionException xpee) {
+                xPathEvaluator = new XPathEvaluator(matcher, namespacePrefixes);
+            } catch (Throwable throwable) {
                 if (MockServerLogger.isEnabled(DEBUG)) {
                     mockServerLogger.logEvent(
                         new LogEntry()
                             .setLogLevel(DEBUG)
                             .setMessageFormat("error while creating xpath expression for{}assuming matcher not xpath{}")
-                            .setArguments(matcher, xpee.getMessage())
-                            .setThrowable(xpee)
+                            .setArguments(matcher, throwable.getMessage())
+                            .setThrowable(throwable)
                     );
                 }
             }
@@ -81,7 +61,7 @@ public class XPathMatcher extends BodyMatcher<String> {
         boolean result = false;
         boolean alreadyLoggedMatchFailure = false;
 
-        if (xpathExpression == null) {
+        if (xPathEvaluator == null) {
             if (context != null) {
                 context.addDifference(mockServerLogger, "xpath match failed expected:{}found:{}failed because:{}", "null", matched, "xpath matcher was null");
                 alreadyLoggedMatchFailure = true;
@@ -90,11 +70,11 @@ public class XPathMatcher extends BodyMatcher<String> {
             result = true;
         } else if (matched != null) {
             try {
-                result = (Boolean) xpathExpression.evaluate(stringToXmlDocumentParser.buildDocument(matched, (matchedInException, throwable, level) -> {
+                result = (Boolean) xPathEvaluator.evaluateXPathExpression(matched, (matchedInException, throwable, level) -> {
                     if (context != null) {
                         context.addDifference(mockServerLogger, throwable, "xpath match failed expected:{}found:{}failed because " + prettyPrint(level) + ":{}", matcher, matched, throwable.getMessage());
                     }
-                }, namespaceAware), XPathConstants.BOOLEAN);
+                }, XPathConstants.BOOLEAN);
             } catch (Throwable throwable) {
                 if (context != null) {
                     context.addDifference(mockServerLogger, throwable, "xpath match failed expected:{}found:{}failed because:{}", matcher, matched, throwable.getMessage());